Fix pointer storage for interfaces
Converted localvar->make_shared to the correct shared_ptr from outset.
Same functionality, but will actually work with calling the function
pointers and associated details.

In the event of a library not exporting the relevant functions, the
shared pointer will be deleted on the next loop due to scope exit.

Signed-off-by: James Warren <admin@trezanik.org>